requirements.txt
apache-airflow==1.10.3
uwsgi==2.0.15